Refactor platform so that it splits moving and and sensor-holding functionality. Removes weakref cycles
