+++ docs/examples/Sensor_Platform_Simulation.py
@@ -174,11 +174,11 @@ platform.sensors
 # platform functionality which is shown in a subsequent example.
 #
 # We can also check to ensure that the default mounting_offsets have been applied:
-platform.mounting_offsets
+radar.mounting_offset
 
 # %%
 # And that the rotation_offsets have been applied:
-platform.rotation_offsets
+radar.rotation_offset
 
 # %%
 # Building a simulation
