+++ stonesoup/platform/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
-from .base import Platform
+from .base import Platform, MovingPlatform, FixedPlatform, MultiTransitionMovingPlatform
 
-__all__ = ['Platform']
+__all__ = ['Platform', 'MovingPlatform', 'FixedPlatform', 'MultiTransitionMovingPlatform']
