+++ stonesoup/sensor/base.py
@@ -1,79 +1,79 @@
 # -*- coding: utf-8 -*-
-import weakref
-from abc import abstractmethod, ABC
-from typing import Optional, Set, Union
-from warnings import warn
+from abc import ABC
+from typing import Optional
 
-import numpy as np
+from stonesoup.moveable.movable import Movable, FixedMovable
+from stonesoup.types.state import State
 
-from ..base import Base
-from ..platform import Platform
+from ..base import Base, Property
 from ..types.array import StateVector
-from ..types.detection import TrueDetection
-from ..types.groundtruth import GroundTruthState
 
 
-class BaseSensor(Base, ABC):
-    """Sensor base class
+class PlatformMountable(Base, ABC):
+    """Base class for any objects that is part of a platform.
 
-    .. warning::
-        This class is private and should not be used or subclassed directly. Instead use the
-        :class:`~.Sensor` class which is needed to achieve the functionality described in this
-        class's documentation.
+    All PlatformMountables must be mounted on a platform to calculate their position and
+    orientation. To make this easier, if the sensor has a position and/or orientation specified in
+    the constructor, and no :attr:`platform_system`, then the default is to create an internally
+    held "private" platform for the Sensor. This allows the Sensor to control (and set) its own
+    position and orientation.
 
     """
 
+    rotation_offset: StateVector = Property(
+        default=None,
+        doc="A StateVector containing the sensor rotation "
+            "offsets from the platform's primary axis (defined as the "
+            "direction of motion). Defaults to a zero vector with the "
+            "same length as the Platform's :attr:`position_mapping`")
+    mounting_offset: StateVector = Property(
+        default=None,
+        doc="A StateVector containing the sensor translation "
+            "offsets from the platform's reference point. Defaults to "
+            "a zero vector with the same length as the Platform's "
+            ":attr:`position_mapping`")
+
+    movement_controller: Movable = Property(
+        default=None,
+        doc="The :class:`~.`Movable` object that controls the movement of this sensor. Will be "
+            "set by the platform if the sensor is assigned to a platform."
+    )
+
     def __init__(self, *args, **kwargs):
+        position = kwargs.pop('position', None)
+        orientation = kwargs.pop('orientation', None)
+        self._internal_movement_controller = None
         super().__init__(*args, **kwargs)
-        self._platform_system = None
+        if position is not None or orientation is not None:
+            if position is None:
+                # assuming 3d for a default platform
+                position = StateVector([0, 0, 0])
+            if orientation is None:
+                orientation = StateVector([0, 0, 0])
+            self._internal_movement_controller = FixedMovable(
+                states=State(state_vector=position),
+                position_mapping=list(range(len(position))),
+                orientation=orientation)
+            self._property_movement_controller = self._internal_movement_controller
+            self._set_mounting_rotation_defaults()
 
-    @property
-    def platform(self) -> Optional[Platform]:
-        """Return the platform system to which the sensor is attached. Resolves the ``weakref``
-        stored in the :attr:`platform_system` Property."""
-        if self.platform_system is None:
-            return None
-        else:
-            return self.platform_system()
+    def _set_mounting_rotation_defaults(self):
+        if self.movement_controller is None:
+            return
+        if self.mounting_offset is None:
+            self.mounting_offset = (StateVector([0]
+                                    * len(self.movement_controller.position_mapping)))
 
-    @property
-    def platform_system(self) -> Optional[weakref.ref]:
-        """Return a ``weakref`` to the platform on which the sensor is mounted"""
-        return self._platform_system
+        if self.rotation_offset is None:
+            self.rotation_offset = StateVector([0] * 3)
 
-    @platform_system.setter
-    def platform_system(self, value: weakref.ref):
-        # this slightly odd construction is to allow overriding by the Sensor subclass
-        self._set_platform_system(value)
-
-    def _set_platform_system(self, value: weakref.ref):
-        if self._platform_system is not None:
-            warn('Sensor has been moved from one platform to another. This is unexpected '
-                 'behaviour')
-        self._platform_system = value
-
-    @abstractmethod
-    def measure(self, ground_truths: Set[GroundTruthState], noise: Union[np.ndarray, bool] = True,
-                **kwargs) -> Set[TrueDetection]:
-        """Generate a measurement for a given state
-
-        Parameters
-        ----------
-        ground_truths : Set[:class:`~.GroundTruthState`]
-            A set of :class:`~.GroundTruthState`
-        noise: :class:`numpy.ndarray` or bool
-            An externally generated random process noise sample (the default is `True`, in which
-            case :meth:`~.Model.rvs` is used; if `False`, no noise will be added)
-
-        Returns
-        -------
-        Set[:class:`~.TrueDetection`]
-            A set of measurements generated from the given states. The timestamps of the
-            measurements are set equal to that of the corresponding states that they were
-            calculated from. Each measurement stores the ground truth path that it was produced
-            from.
-        """
-        raise NotImplementedError
+    @movement_controller.setter
+    def movement_controller(self, value):
+        if self._has_internal_controller:
+            raise AttributeError('The movement controller cannot be set on a Sensor with a private'
+                                 'internal controller')
+        self._property_movement_controller = value
+        self._set_mounting_rotation_defaults()
 
     @property
     def position(self) -> Optional[StateVector]:
@@ -81,21 +81,23 @@ class BaseSensor(Base, ABC):
         of Cartesian coordinates in the order :math:`x,y,z`.
 
         .. note::
-            This property delegates the actual calculation of position to the platform on which the
-            sensor is mounted.
+            This property delegates the actual calculation of position to the Sensor's
+            :attr:`movement_controller`
 
-            It is settable if, and only if, the sensor holds its own internal platform."""
-        if self.platform is None:
+            It is settable if, and only if, the sensor holds its own internal movement_controller.
+            """
+        if self.movement_controller is None:
             return None
-        return self.platform.get_sensor_position(self)
+        return (self.movement_controller.position
+                + self.movement_controller._get_rotated_offset(self.mounting_offset))
 
     @position.setter
     def position(self, value: StateVector):
-        if self._has_internal_platform:
-            self.platform.position = value
+        if self._has_internal_controller:
+            self.movement_controller.position = value
         else:
             raise AttributeError('Cannot set sensor position unless the sensor has its own '
-                                 'default platform')
+                                 'default movement_controller')
 
     @property
     def orientation(self) -> Optional[StateVector]:
@@ -106,25 +108,26 @@ class BaseSensor(Base, ABC):
         towards the origin.
 
         .. note::
-            This property delegates the actual calculation of orientation to the platform on which
-            the sensor is mounted.
+            This property delegates the actual calculation of orientation to the Sensor's
+            :attr:`movement_controller`
 
-            It is settable if, and only if, the sensor holds its own internal platform."""
-        if self.platform is None:
+            It is settable if, and only if, the sensor holds its own internal movement_controller.
+            """
+        if self.movement_controller is None:
             return None
-        return self.platform.get_sensor_orientation(self)
+        return self.movement_controller.orientation + self.rotation_offset
 
     @orientation.setter
     def orientation(self, value: StateVector):
-        if self._has_internal_platform:
-            self.platform.orientation = value
+        if self._has_internal_controller:
+            self.movement_controller.orientation = value
         else:
             raise AttributeError('Cannot set sensor position unless the sensor has its own '
-                                 'default platform')
+                                 'default movement_controller')
 
     @property
-    def _has_internal_platform(self) -> bool:
-        return False
+    def _has_internal_controller(self):
+        return self._internal_movement_controller is not None
 
     @property
     def velocity(self) -> Optional[StateVector]:
@@ -132,9 +135,9 @@ class BaseSensor(Base, ABC):
         of Cartesian coordinates in the order :math:`x,y,z`.
 
         .. note::
-            This property delegates the actual calculation of velocity to the platform on which the
-            sensor is mounted.
+            This property delegates the actual calculation of velocity to the Sensor's
+            :attr:`movement_controller`
 
-            It is settable if, and only if, the sensor holds its own internal platform which is
-            a MovingPlatfom."""
-        return self.platform.velocity
+            It is settable if, and only if, the sensor holds its own internal movement_controller
+            which is a :class:`~.MovingMovable`."""
+        return self.movement_controller.velocity
